requirements.txt
absl-py==1.4.0
aes==1.2.0
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.2
annotated-types==0.6.0
ansi2html==1.8.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
auto-py-to-exe==2.23.0
autogui==0.1.8
Automat==22.10.0
backcall==0.2.0
Bashutils==0.0.4
bcrypt==3.2.2
beautifulsoup4==4.11.1
biopython==1.81
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cleo==0.8.1
click==8.1.3
clikit==0.6.2
clr-loader==0.2.4
cmdstanpy==1.2.2
colorama==0.4.4
comtypes==1.1.14
config==0.5.1
configparser==7.0.0
constantly==15.1.0
crapy==0.0.2
crashtest==0.3.1
crayons==0.4.0
crypto==1.4.1
cryptography==36.0.2
cssselect==1.2.0
Custom-Library==0.0.2
cvxpy==1.3.2
cx-Freeze==6.11.1
cycler==0.11.0
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
distro==1.8.0
Django==5.0.3
django-extensions==3.2.3
djangorestframework==3.15.1
docopt==0.6.2
dotty-dict==1.3.1
ecos==2.0.12
Eel==0.14.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
exceptionite==2.2.5
executing==0.8.3
Faker==13.16.0
fastjsonschema==2.18.1
feedparser==6.0.11
filelock==3.10.0
findspark==2.0.1
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.37.1
frozendict==2.3.8
frozenlist==1.4.0
future==0.18.2
gast==0.4.0
gevent==21.12.0
gevent-websocket==0.10.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==1.1.3
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==22.0.0
h11==0.13.0
h5py==3.11.0
hashids==1.3.1
hfilesize==0.1.0
holidays==0.47
html5lib==1.1
httpcore==1.0.2
httpx==0.25.1
hupper==1.10.3
hyperlink==21.0.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.8.0
importlib_resources==6.4.0
incremental==22.10.0
inflection==0.3.1
iniconfig==2.0.0
ipykernel==6.13.1
ipython==8.4.0
ipywidgets==8.0.4
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.2.0
Js2Py==0.74
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-widgets==3.0.5
keras==3.3.2
keyboard==0.13.5
kiwisolver==1.4.4
layers==0.1.5
lazy_loader==0.2
libclang==16.0.0
lief==0.12.1
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
masonite==4.18.0
masonite-orm==2.20.0
matplotlib==3.5.3
matplotlib-inline==0.1.3
mdurl==0.1.2
ml-dtypes==0.3.2
mlxtend==0.23.1
mock==5.1.0
MouseInfo==0.1.3
mpmath==1.2.1
multidict==6.0.4
multitasking==0.0.11
mysqlclient==2.2.4
Naked==0.1.32
namex==0.0.8
nbformat==5.9.2
nest-asyncio==1.5.5
networkx==3.2.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
osqp==0.6.3
outcome==1.2.0
packaging==21.3
pandas==1.4.2
parsel==1.7.0
parso==0.8.3
pastel==0.2.1
patsy==0.5.3
pefile==2022.5.30
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.2.0
pipwin==0.5.2
plotly==5.16.1
pluggy==1.3.0
prettytable==3.3.0
progressbar2==2.7.3
prompt-toolkit==3.0.29
prophet==1.1.5
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PuLP==2.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/pc/pipwin/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyDispatcher==2.0.7
pygame==2.1.2
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyjsparser==2.7.1
PyJWT==2.4.0
pylev==1.4.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pynput==1.7.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyportfolioopt==1.5.5
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
pyspark==3.4.0
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==0.15.0
pythonnet==3.0.1
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyWavelets==1.4.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==23.1.0
qdldl==0.1.7.post0
queuelib==1.6.2
raise==1.1.9
referencing==0.30.2
regex==2023.3.23
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.1
rpds-py==0.10.6
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.11.3
Scrapy==2.8.0
scs==3.2.3
seaborn==0.12.2
selenium==4.20.0
sequential==1.0.0
service-identity==21.1.0
sgmllib3k==1.0.0
shellescape==3.8.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
srapy==0.1.19
stack-data==0.2.0
stanio==0.5.0
statsmodels==0.14.0
sympy==1.11.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.14.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tldextract==2.2.3
tomli==2.0.1
tornado==6.1
tqdm==4.64.1
traitlets==5.2.2.post1
triggers==0.0.3
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.11.0
tzdata==2022.7
tzlocal==4.2
ucimlrepo==0.0.3
urllib3==1.26.12
w3lib==2.1.1
watchdog==2.3.1
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==2.2.3
whichcraft==0.6.1
whitenoise==5.2.0
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
yarl==1.9.2
yfinance==0.2.28
zipp==3.17.0
zope.event==4.5.0
zope.interface==5.4.0
